fix: keep Windows PID guards portable

+++ internal/cli/process_windows.go
@@ -21,7 +21,7 @@ var (
 func processRunning(pid int) (bool, error) {
 	// Windows OpenProcess takes a DWORD (uint32). Reject non-positive values and
 	// anything that would truncate on the architecture-dependent int → uint32 cast.
-	if pid <= 0 || pid > math.MaxUint32 {
+	if pid <= 0 || int64(pid) > int64(math.MaxUint32) {
 		return false, nil
 	}
 	handle, err := windows.OpenProcess(windows.SYNCHRONIZE|windows.PROCESS_QUERY_LIMITED_INFORMATION, false, uint32(pid))
@@ -43,7 +43,7 @@ func processRunning(pid int) (bool, error) {
 }
 
 func processIdentityMatches(pid int, expected string) (bool, error) {
-	if pid <= 0 || pid > math.MaxUint32 {
+	if pid <= 0 || int64(pid) > int64(math.MaxUint32) {
 		return false, nil
 	}
 	if strings.TrimSpace(expected) == "" {
@@ -67,7 +67,7 @@ func processIdentityMatches(pid int, expected string) (bool, error) {
 }
 
 func stopProcess(pid int, expected ...string) error {
-	if pid <= 0 || pid > math.MaxUint32 {
+	if pid <= 0 || int64(pid) > int64(math.MaxUint32) {
 		return fmt.Errorf("invalid process pid %d", pid)
 	}
 	identity := ""
